+++ src/Deriver/Types.fs
@@ -20,13 +20,13 @@ type DeriverApplicationError =
 
 type DeriveEvent<'InputEvent, 'OutputEvent> = 'InputEvent -> 'OutputEvent list
 
-type DeriveEventHandler<'InputEvent, 'OutputEvent> =
+type internal DeriveEventHandler<'InputEvent, 'OutputEvent> =
     | Simple of DeriveEvent<'InputEvent, 'OutputEvent>
     | WithApplication of (PatternRuntimeParts -> DeriveEvent<'InputEvent, 'OutputEvent>)
 
 // Deriver Application Configuration
 
-type DeriverParts<'InputEvent, 'OutputEvent> = {
+type internal DeriverParts<'InputEvent, 'OutputEvent> = {
     Configuration: Configuration<'InputEvent, 'OutputEvent> option
     DeriveTo: ConnectionName option
     DeriveEvent: DeriveEventHandler<'InputEvent, 'OutputEvent> option
@@ -34,7 +34,7 @@ type DeriverParts<'InputEvent, 'OutputEvent> = {
     GetCommonEvent: GetCommonEvent<'InputEvent, 'OutputEvent> option
 }
 
-module DeriverParts =
+module internal DeriverParts =
     let defaultDeriver = {
         Configuration = None
         DeriveTo = None
@@ -51,5 +51,6 @@ type DeriverApplicationParts<'InputEvent, 'OutputEvent> = {
 
 type DeriverApplication<'InputEvent, 'OutputEvent> = internal DeriverApplication of Result<DeriverApplicationParts<'InputEvent, 'OutputEvent>, DeriverApplicationError>
 
-module DeriverApplication =
+[<RequireQualifiedAccess>]
+module internal DeriverApplication =
     let application { Application = application } = application